+++ src/Halcyon.Api/Core/Authentication/IJwtTokenGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using Halcyon.Api.Data;
 
-namespace Halcyon.Api.Services.Jwt;
+namespace Halcyon.Api.Core.Authentication;
 
 public interface IJwtTokenGenerator
 {
