+++ src/Halcyon.Api/Core/Email/IEmailSender.cs
@@ -1,4 +1,4 @@
-﻿namespace Halcyon.Api.Services.Email;
+﻿namespace Halcyon.Api.Core.Email;
 
 public interface IEmailSender
 {
