+++ src/Halcyon.Api/Core/Web/CorsPolicySettings.cs
@@ -1,12 +1,12 @@
-﻿namespace Halcyon.Api.Common;
+﻿namespace Halcyon.Api.Core.Web;
 
 public class CorsPolicySettings
 {
     public static string SectionName { get; } = "CorsPolicy";
 
     public string[] AllowedOrigins { get; set; }
-    
+
     public string[] AllowedMethods { get; set; }
- 
+
     public string[] AllowedHeaders { get; set; }
 }
