+++ src/Halcyon.Api/Core/Web/CurrentUser.cs
@@ -1,4 +1,4 @@
-﻿namespace Halcyon.Api.Common;
+﻿namespace Halcyon.Api.Core.Web;
 
 public record CurrentUser(Guid Id)
 {
