+++ src/Halcyon.Api/Core/Web/IEndpoint.cs
@@ -1,4 +1,4 @@
-﻿namespace Halcyon.Api.Common;
+﻿namespace Halcyon.Api.Core.Web;
 
 public interface IEndpoint
 {
