+++ src/Halcyon.Api/Core/Web/UpdateResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace Halcyon.Api.Common;
+﻿namespace Halcyon.Api.Core.Web;
 
 public class UpdateResponse
 {
