+++ src/Halcyon.Api/Core/Web/ValidationFilter.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace Halcyon.Api.Common;
+namespace Halcyon.Api.Core.Web;
 
 public class ValidationFilter(IServiceProvider serviceProvider) : IEndpointFilter
 {
