+++ src/Halcyon.Api/Data/Role.cs
@@ -3,5 +3,6 @@
 public static class Role
 {
     public const string SystemAdministrator = "SYSTEM_ADMINISTRATOR";
+
     public const string UserAdministrator = "USER_ADMINISTRATOR";
 }
