+++ src/Halcyon.Api/Features/Manage/UpdateProfile/UpdateProfileEndpoint.cs
@@ -1,4 +1,4 @@
-﻿using Halcyon.Api.Common;
+﻿using Halcyon.Api.Core.Web;
 using Halcyon.Api.Data;
 using Mapster;
 using Microsoft.EntityFrameworkCore;
