+++ src/Halcyon.Api/Features/Messaging/MessageConsumer.cs
@@ -1,5 +1,6 @@
 using Halcyon.Api.Data;
 using Halcyon.Api.Features.Messaging;
+using Halcyon.Api.Core.Database;
 using MassTransit;
 using Microsoft.AspNetCore.SignalR;
