+++ src/Halcyon.Api/Features/Token/TokenEndpoint.cs
@@ -1,7 +1,6 @@
-﻿using Halcyon.Api.Common;
+﻿using Halcyon.Api.Core.Authentication;
+using Halcyon.Api.Core.Web;
 using Halcyon.Api.Data;
-using Halcyon.Api.Services.Hash;
-using Halcyon.Api.Services.Jwt;
 using Microsoft.EntityFrameworkCore;
 
 namespace Halcyon.Api.Features.Token;
