+++ src/Halcyon.Api/Features/Users/SearchUsers/SearchUsersEndpoint.cs
@@ -1,4 +1,4 @@
-﻿using Halcyon.Api.Common;
+﻿using Halcyon.Api.Core.Web;
 using Halcyon.Api.Data;
 using Mapster;
 using Microsoft.EntityFrameworkCore;
