+++ test/Halcyon.Api.Tests/Features/Account/Register/RegisterEndpointTests.cs
@@ -1,17 +1,14 @@
 using System.Net;
 using System.Net.Http.Json;
-using Halcyon.Api.Common;
+using Halcyon.Api.Core.Web;
 using Halcyon.Api.Features.Account.Register;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Halcyon.Api.Tests.Features.Account.Register;
 
-public class RegisterEndpointTests : BaseTest
+public class RegisterEndpointTests(TestWebApplicationFactory factory) : BaseTest(factory)
 {
     private const string _requestUri = "/account/register";
-
-    public RegisterEndpointTests(TestWebApplicationFactory factory)
-        : base(factory) { }
 
     [Fact]
     public async Task Register_ShouldReturnBadRequest_WhenDuplicateEmailAddress()
